fix: fix wallet connect flow
